+++ src/gtk/gtk3/NautilusColumnProvider.c
@@ -51,6 +51,7 @@ const struct rp_nautilus_column_provider_column_desc_data_t rp_nautilus_column_p
 	{"rp-region-code",	"Region Code"},
 	{"rp-category",		"Category"},
 	{"rp-ripped-by",	"Ripped By"},
+	{"rp-version",		"Version"},
 
 	{NULL,			NULL}
 };