+++ src/win32/RP_ContextMenu.cpp
@@ -91,14 +91,14 @@ int RP_ContextMenu_Private::convert_to_png(LPCTSTR source_filename)
 	output_filename.assign(source_filename, source_filename_len);
 
 	// Find the current extension and replace it.
-	const size_t dotpos = output_filename.find_last_of(_T('.'));
+	const size_t dotpos = output_filename.rfind(_T('.'));
 	if (dotpos == tstring::npos) {
 		// No file extension. Add it.
 		output_filename += _T(".png");
 	} else {
 		// If the dot is after the last slash, we already have a file extension.
 		// Otherwise, we don't have one, and need to add it.
-		const size_t slashpos = output_filename.find_last_of(_T('\\'));
+		const size_t slashpos = output_filename.rfind(_T('\\'));
 		if (slashpos < dotpos) {
 			// We already have a file extension. Remove it first.
 			output_filename.resize(dotpos);